Extract MessageCaster and implementations into package visibility as TypeConverter is too overloaded.
