+++ base/src/main/java/io/spine/protobuf/BytesCaster.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2020, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.protobuf;
+
+import com.google.protobuf.ByteString;
+import com.google.protobuf.BytesValue;
+
+/**
+ * Converts {@link BytesValue} to {@link ByteString} and back.
+ */
+final class BytesCaster extends MessageCaster<BytesValue, ByteString> {
+
+    @Override
+    protected ByteString toObject(BytesValue input) {
+        ByteString result = input.getValue();
+        return result;
+    }
+
+    @Override
+    protected BytesValue toMessage(ByteString input) {
+        BytesValue bytes = BytesValue
+                .newBuilder()
+                .setValue(input)
+                .build();
+        return bytes;
+    }
+}
